extracted renderer to MainRenderer class
